@@ -1,11 +1,9 @@
 'use strict';
 
-var express = require('express');
-var controller = require('./user.controller');
-var config = require('../../config/environment');
-var auth = require('../../auth/auth.service');
-
-var router = express.Router();
+import {Router} from 'express';
+import * as controller from './user.controller';
+import * as auth from '../../auth/auth.service';
+var router = new Router();
 
 router.get('/',  controller.list);
 router.get('/past', controller.past);
@@ -34,4 +32,4 @@ router.put("/:id/addTech", auth.canEdit(), controller.addTech);
 router.put("/:id/removeTech", auth.canEdit(), controller.removeTech);
 router.post('/resetPassword', controller.resetPassword);
 
-module.exports = router;
+export default router;
