@@ -1,64 +1,74 @@
 'use strict';
 
-var should = require('should');
-var app = require('../../app');
-var User = require('./user.model');
+import app from '../..';
+import User from './user.model';
+var user;
+var genUser = function() {
+  user = new User({
+    provider: 'local',
+    name: 'Fake User',
+    email: 'test@example.com',
+    password: 'password',
+    avatar: '//www.gravatar.com/avatar/00000000000000000000000000000000?d=monsterid'
 
-var user = new User({
-  provider: 'local',
-  name: 'Fake User',
-  email: 'test@test.com',
-  password: 'password',
-  avatar: '//www.gravatar.com/avatar/00000000000000000000000000000000?d=monsterid'
-});
+  });
+  return user;
+};
 
 describe('User Model', function() {
-  before(function(done) {
+  before(function() {
     // Clear users before testing
-    User.remove().exec().then(function() {
-      done();
+    return User.removeAsync();
+  });
+
+  beforeEach(function() {
+    genUser();
+  });
+
+  afterEach(function() {
+    return User.removeAsync();
+  });
+
+  it('should begin with no users', function() {
+    return expect(User.findAsync({})).to
+      .eventually.have.length(0);
+  });
+
+  it('should fail when saving a duplicate user', function() {
+    return expect(user.saveAsync()
+      .then(function() {
+        var userDup = genUser();
+        return userDup.saveAsync();
+      })).to.be.rejected;
+  });
+
+  describe('#email', function() {
+    it('should fail when saving without an email', function() {
+      user.email = '';
+      return expect(user.saveAsync()).to.be.rejected;
     });
   });
 
-  afterEach(function(done) {
-    User.remove().exec().then(function() {
-      done();
+  describe('#password', function() {
+    beforeEach(function() {
+      return user.saveAsync();
+    });
+
+    it('should authenticate user if valid', function() {
+      expect(user.authenticate('password')).to.be.true;
+    });
+
+    it('should not authenticate user if invalid', function() {
+      expect(user.authenticate('blah')).to.not.be.true;
+    });
+
+    it('should remain the same hash unless the password is updated', function() {
+      user.name = 'Test User';
+      return expect(user.saveAsync()
+        .spread(function(u) {
+          return u.authenticate('password');
+        })).to.eventually.be.true;
     });
   });
 
-  it('should begin with no users', function(done) {
-    User.find({}, function(err, users) {
-      users.should.have.length(0);
-      done();
-    });
-  });
-
-  it('should fail when saving a duplicate user', function(done) {
-    user.save(function() {
-      var userDup = new User(user);
-      userDup.save(function(err) {
-        should.exist(err);
-        done();
-      });
-    });
-  });
-
-  it('should fail when saving without an email', function(done) {
-    user.email = '';
-    user.save(function(err) {
-      should.exist(err);
-      done();
-    });
-  });
-
-  it("should authenticate user if password is valid", function() {
-    return user.authenticate('password').should.be.true;
-  });
-
-  it("should not authenticate user if password is invalid", function() {
-    return user.authenticate('blah').should.not.be.true;
-  });
- it("should not authenticate user if password is blank", function() {
-    return user.authenticate('').should.not.be.true;
-  });
 });
