@@ -1,15 +1,19 @@
 'use strict';
 
-var express = require('express');
-var passport = require('passport');
-var config = require('../config/environment');
-var User = require('../api/user/user.model');
+import express from 'express';
+import passport from 'passport';
+import config from '../config/environment';
+import User from '../api/user/user.model';
 
 // Passport Configuration
 require('./local/passport').setup(User, config);
-
+// require('./facebook/passport').setup(User, config);
+// require('./google/passport').setup(User, config);
+//
 var router = express.Router();
 
 router.use('/local', require('./local'));
-
-module.exports = router;
+// router.use('/facebook', require('./facebook'));
+// router.use('/google', require('./google'));
+//
+export default router;
