@@ -1,20 +1,24 @@
 'use strict';
 
-var express = require('express');
-var passport = require('passport');
-var auth = require('../auth.service');
+import express from 'express';
+import passport from 'passport';
+import {signToken} from '../auth.service';
 var User = require("../../api/user/user.model");
 
 var router = express.Router();
 
 router.post('/', function(req, res, next) {
-  passport.authenticate('local', function (err, user, info) {
+  passport.authenticate('local', function(err, user, info) {
     var error = err || info;
-    if (error) return res.json(401, error);
-    if (!user) return res.json(404, {message: 'Something went wrong, please try again.'});
+    if (error) {
+      return res.status(401).json(error);
+    }
+    if (!user) {
+      return res.status(404).json({message: 'Something went wrong, please try again.'});
+    }
 
-    var token = auth.signToken(user._id, user.role);
-    res.json({token: token});
+    var token = signToken(user._id, user.role);
+    res.json({ token });
   })(req, res, next)
 });
 
@@ -31,8 +35,8 @@ router.post('/token', function(req, res, next){
         return;
     }
 
-		if (new Date() < user.passwordResetExpiration){
-      var token = auth.signToken(user._id, user.role);
+	if (new Date() < user.passwordResetExpiration){
+    var token = signToken(user._id, user.role);
       res.json({token: token});
     }else{
       // TODO make a page
@@ -41,4 +45,4 @@ router.post('/token', function(req, res, next){
 	});
 });
 
-module.exports = router;
+export default router;
