@@ -1,25 +1,35 @@
-var passport = require('passport');
-var LocalStrategy = require('passport-local').Strategy;
+import passport from 'passport';
+import {Strategy as LocalStrategy} from 'passport-local';
 
-exports.setup = function (User, config) {
-  passport.use(new LocalStrategy({
-      usernameField: 'email',
-      passwordField: 'password' // this is the virtual field on the model
-    },
-    function(email, password, done) {
-      User.findOne({
-        email: email.toLowerCase()
-      }, function(err, user) {
-        if (err) return done(err);
-
-        if (!user) {
-          return done(null, false, { message: 'This email is not registered.' });
+function localAuthenticate(User, email, password, done) {
+  User.findOneAsync({
+    email: email.toLowerCase()
+  })
+    .then(user => {
+      if (!user) {
+        return done(null, false, {
+          message: 'This email is not registered.'
+        });
+      }
+      user.authenticate(password, function(authError, authenticated) {
+        if (authError) {
+          return done(authError);
         }
-        if (!user.authenticate(password)) {
+        if (!authenticated) {
           return done(null, false, { message: 'This password is not correct.' });
+        } else {
+          return done(null, user);
         }
-        return done(null, user);
       });
-    }
-  ));
-};
+    })
+    .catch(err => done(err));
+}
+
+export function setup(User, config) {
+  passport.use(new LocalStrategy({
+    usernameField: 'email',
+    passwordField: 'password' // this is the virtual field on the model
+  }, function(email, password, done) {
+    return localAuthenticate(User, email, password, done);
+  }));
+}
