@@ -1,22 +1,37 @@
 'use strict';
 
 angular.module('observatory3App')
-  .config(function ($stateProvider) {
+  .config(function($stateProvider) {
     $stateProvider
       .state('login', {
         url: '/login',
         templateUrl: 'app/account/login/login.html',
-        controller: 'LoginCtrl'
+        controller: 'LoginController',
+        controllerAs: 'vm'
+      })
+      .state('logout', {
+        url: '/logout?referrer',
+        referrer: 'main',
+        template: '',
+        controller: function($state, Auth) {
+          var referrer = $state.params.referrer ||
+                          $state.current.referrer ||
+                          'main';
+          Auth.logout();
+          $state.go(referrer);
+        }
       })
       .state('signup', {
         url: '/signup',
         templateUrl: 'app/account/signup/signup.html',
-        controller: 'SignupCtrl'
+        controller: 'SignupController',
+        controllerAs: 'vm'
       })
       .state('settings', {
         url: '/settings',
         templateUrl: 'app/account/settings/settings.html',
         controller: 'SettingsCtrl',
+        controllerAs: 'vm',
         authenticate: true
       })
       .state('forgot-password', {
@@ -24,4 +39,11 @@ angular.module('observatory3App')
         templateUrl: 'app/account/forgot-password/forgot-password.html',
         controller: 'ForgotPasswordCtrl'
       });
+  })
+  .run(function($rootScope) {
+    $rootScope.$on('$stateChangeStart', function(event, next, nextParams, current) {
+      if (next.name === 'logout' && current && current.name && !current.authenticate) {
+        next.referrer = current.name;
+      }
+    });
   });
