@@ -1,7 +1,7 @@
 'use strict';
 
 angular.module('observatory3App')
-  .controller('LoginCtrl', function ($scope, Auth, $location) {
+  .controller('LoginController', function ($scope, Auth, $location) {
     $scope.user = {};
     $scope.errors = {};
 
