@@ -1,9 +1,11 @@
 'use strict';
-
-angular.module('observatory3App')
-  .factory('Auth', function Auth($location, $rootScope, $http, User, $cookieStore, $q) {
+(function() {
+function AuthService($location, $http, $cookies, $q, appConfig, Util, User) {
     var currentUser = {};
-    if($cookieStore.get('token')) {
+    var safeCb = Util.safeCb;
+    var currentUser = {};
+    var userRoles = appConfig.userRoles || [];
+    if ($cookies.get('token') && $location.path() !== '/logout') {
       currentUser = User.get();
     }
 
@@ -20,38 +22,7 @@ angular.module('observatory3App')
         cb(false);
       }
     }
-
-    return {
-
-      /**
-       * Authenticate user and save token
-       *
-       * @param  {Object}   user     - login info
-       * @param  {Function} callback - optional
-       * @return {Promise}
-       */
-      login: function(user, callback) {
-        var cb = callback || angular.noop;
-        var deferred = $q.defer();
-
-        $http.post('/auth/local', {
-          email: user.email,
-          password: user.password
-        }).
-        success(function(data) {
-          $cookieStore.put('token', data.token);
-          currentUser = User.get();
-          deferred.resolve(data);
-          return cb();
-        }).
-        error(function(err) {
-          this.logout();
-          deferred.reject(err);
-          return cb(err);
-        }.bind(this));
-
-        return deferred.promise;
-      },
+  var Auth = {
 
       /**
        * Authenticate user via reset token and save session token
@@ -61,171 +32,236 @@ angular.module('observatory3App')
        * @return {Promise}
        */
       loginWithResetToken: function(resetToken, callback) {
-        var cb = callback || angular.noop;
-        var deferred = $q.defer();
-
-        $http.post('/auth/local/token', {
+        return $http.post('/auth/local/token', {
           token: resetToken
-        }).
-        success(function(data) {
-          $cookieStore.put('token', data.token);
-          currentUser = User.get();
-          deferred.resolve(data);
-          return cb();
-        }).
-        error(function(err) {
-          this.logout();
-          deferred.reject(err);
-          return cb(err);
-        }.bind(this));
-
-        return deferred.promise;
-      },
-
-      /**
-       * Delete access token and user info
-       *
-       * @param  {Function}
-       */
-      logout: function() {
-        $cookieStore.remove('token');
-        currentUser = {};
-      },
-
-      /**
-       * Create a new user
-       *
-       * @param  {Object}   user     - user info
-       * @param  {Function} callback - optional
-       * @return {Promise}
-       */
-      createUser: function(user, callback) {
-        var cb = callback || angular.noop;
-
-        return User.save(user,
-          function(data) {
-            $cookieStore.put('token', data.token);
+        })
+          .then(res => {
+            $cookies.put('token', res.data.token);
             currentUser = User.get();
-            return cb(user);
-          },
-          function(err) {
-            this.logout();
-            return cb(err);
-          }.bind(this)).$promise;
-      },
-
-      /**
-       * Change password
-       *
-       * @param  {String}   oldPassword
-       * @param  {String}   newPassword
-       * @param  {Function} callback    - optional
-       * @return {Promise}
-       */
-      changePassword: function(oldPassword, newPassword, callback) {
-        var cb = callback || angular.noop;
-
-        return User.changePassword({ id: currentUser._id }, {
-          oldPassword: oldPassword,
-          newPassword: newPassword
-        }, function(user) {
-          return cb(user);
-        }, function(err) {
-          return cb(err);
-        }).$promise;
-      },
-
-      /**
-       * Change password using token
-       *
-       * @param  {String}   token
-       * @param  {String}   newPassword
-       * @param  {Function} callback    - optional
-       * @return {Promise}
-       */
-      changePasswordWithToken: function(token, newPassword, callback) {
-        var cb = callback || angular.noop;
-
-        return User.changePassword({ id: currentUser._id }, {
-          token: token,
-          newPassword: newPassword
-        }, function(user) {
-          return cb(user);
-        }, function(err) {
-          return cb(err);
-        }).$promise;
-      },
-
-      /**
-       * Gets all available info on authenticated user
-       *
-       * @param {Function} OPTIONAL callback
-       * @return {Object} user
-       */
-      getCurrentUser: function(callback) {
-        if (!callback){
-          return currentUser;
-        }else{
-          // Wait for user login status
-          isLoggedInAsync(function(){
-            return callback(currentUser);
+            return currentUser.$promise;
+          })
+          .then(user => {
+            safeCb(callback)(null, user);
+            return user;
+          })
+          .catch(err => {
+            Auth.logout();
+            safeCb(callback)(err.data);
+            return $q.reject(err.data);
           });
-        }
       },
 
-      /**
-       * Check if a user is logged in
-       *
-       * @return {Boolean}
-       */
-      isLoggedIn: function() {
-        return currentUser.hasOwnProperty('role');
-      },
 
-      /**
-       * Waits for currentUser to resolve before checking if user is logged in
-       */
-      isLoggedInAsync: isLoggedInAsync,
-
-      /**
-       * Check if a user is an admin
-       *
-       * @return {Boolean}
-       */
-      isAdmin: function() {
-        return currentUser.role === 'admin';
-      },
-
-      /**
-       * Check if a user is a mentor
-       *
-       * @return {Boolean}
-       */
-      isMentor: function() {
-        return currentUser.role === 'mentor' || currentUser.role === 'admin'; 
-      },
-
-      /**
-       * Get auth token
-       */
-      getToken: function() {
-        return $cookieStore.get('token');
-      },
-
-      /**
-       * Sends password reminder to specified email
-       *
-       * @param {String} email
-       * @return {Promise}
-       **/
-       resetPassword: function(email, callback){
-        $http.post('/api/users/resetPassword', {
-          email: email
-        }).success(function(){
-          callback();
-        }).error(function(err){
-          callback(err);
+    /**
+     * Authenticate user and save token
+     *
+     * @param  {Object}   user     - login info
+     * @param  {Function} callback - optional, function(error, user)
+     * @return {Promise}
+     */
+    login({email, password}, callback) {
+      return $http.post('/auth/local', {
+        email: email,
+        password: password
+      })
+        .then(res => {
+          $cookies.put('token', res.data.token);
+          currentUser = User.get();
+          return currentUser.$promise;
+        })
+        .then(user => {
+          safeCb(callback)(null, user);
+          return user;
+        })
+        .catch(err => {
+          Auth.logout();
+          safeCb(callback)(err.data);
+          return $q.reject(err.data);
         });
-       }
-    };
-  });
+    },
+    /**
+     * Delete access token and user info
+     */
+    logout() {
+      $cookies.remove('token');
+      currentUser = {};
+    },
+    /**
+     * Create a new user
+     *
+     * @param  {Object}   user     - user info
+     * @param  {Function} callback - optional, function(error, user)
+     * @return {Promise}
+     */
+    createUser(user, callback) {
+      return User.save(user,
+        function(data) {
+          $cookies.put('token', data.token);
+          currentUser = User.get();
+          return safeCb(callback)(null, user);
+        },
+        function(err) {
+          Auth.logout();
+          return safeCb(callback)(err);
+        }).$promise;
+    },
+    /**
+     * Change password
+     *
+     * @param  {String}   oldPassword
+     * @param  {String}   newPassword
+     * @param  {Function} callback    - optional, function(error, user)
+     * @return {Promise}
+     */
+    changePassword(oldPassword, newPassword, callback) {
+      return User.changePassword({ id: currentUser._id }, {
+        oldPassword: oldPassword,
+        newPassword: newPassword
+      }, function() {
+        return safeCb(callback)(null);
+      }, function(err) {
+        return safeCb(callback)(err);
+      }).$promise;
+    },
+
+    /**
+    * Change password using token
+    *
+    * @param  {String}   token
+    * @param  {String}   newPassword
+    * @param  {Function} callback    - optional
+    * @return {Promise}
+    */
+    changePasswordWithToken: function(token, newPassword, callback) {
+        return User.changePassword({ id: currentUser._id }, {
+            token: token,
+            newPassword: newPassword
+        }, function() {
+            return safeCb(callback)(null);
+        }, function(err) {
+            return safeCb(callback)(err);
+        }).$promise;
+    },
+
+    /**
+     * Gets all available info on a user
+     *   (synchronous|asynchronous)
+     *
+     * @param  {Function|*} callback - optional, funciton(user)
+     * @return {Object|Promise}
+     */
+    getCurrentUser(callback) {
+      if (arguments.length === 0) {
+        return currentUser;
+      }
+      var value = (currentUser.hasOwnProperty('$promise')) ?
+        currentUser.$promise : currentUser;
+      return $q.when(value)
+        .then(user => {
+          safeCb(callback)(user);
+          return user;
+        }, () => {
+          safeCb(callback)({});
+          return {};
+        });
+    },
+
+    /**
+     * Check if a user is logged in
+     *   (synchronous|asynchronous)
+     *
+     * @param  {Function|*} callback - optional, function(is)
+     * @return {Bool|Promise}
+     */
+    isLoggedIn(callback) {
+      if (arguments.length === 0) {
+        return currentUser.hasOwnProperty('role');
+      }
+      return Auth.getCurrentUser(null)
+        .then(user => {
+          var is = user.hasOwnProperty('role');
+          safeCb(callback)(is);
+          return is;
+        });
+    },
+     /**
+      * Check if a user has a specified role or higher
+      *   (synchronous|asynchronous)
+      *
+      * @param  {String}     role     - the role to check against
+      * @param  {Function|*} callback - optional, function(has)
+      * @return {Bool|Promise}
+      */
+    hasRole(role, callback) {
+      var hasRole = function(r, h) {
+        return userRoles.indexOf(r) >= userRoles.indexOf(h);
+      };
+      if (arguments.length < 2) {
+        return hasRole(currentUser.role, role);
+      }
+      return Auth.getCurrentUser(null)
+        .then(user => {
+          var has = (user.hasOwnProperty('role')) ?
+            hasRole(user.role, role) : false;
+          safeCb(callback)(has);
+          return has;
+        });
+    },
+     /**
+      * Check if a user is an admin
+      *   (synchronous|asynchronous)
+      *
+      * @param  {Function|*} callback - optional, function(is)
+      * @return {Bool|Promise}
+      */
+    isAdmin() {
+      return Auth.hasRole
+        .apply(Auth, [].concat.apply(['admin'], arguments));
+    },
+     /**
+      * Check if a user is an mentor
+      *   (synchronous|asynchronous)
+      *
+      * @param  {Function|*} callback - optional, function(is)
+      * @return {Bool|Promise}
+      */
+    isMentor() {
+      return Auth.hasRole
+        .apply(Auth, [].concat.apply(['mentor'], arguments));
+    },
+
+    /**
+     * Get auth token
+     *
+     * @return {String} - a token string used for authenticating
+     */
+    getToken() {
+      return $cookies.get('token');
+    },
+
+    /**
+    * Waits for currentUser to resolve before checking if user is logged in
+    */
+    isLoggedInAsync: isLoggedInAsync,
+    /**
+    * Sends password reminder to specified email
+    *
+    * @param {String} email
+    * @return {Promise}
+    **/
+    resetPassword(email, callback){
+        return $http.post('/api/users/resetPassword', {
+            email: email
+        }).then(res => {
+          return safeCb(callback)(res);
+        }).catch(err => {
+          return safeCb(callback)(err);
+        });
+    }
+
+  };
+  return Auth;
+}
+angular.module('observatory3App.auth')
+  .factory('Auth', AuthService);
+})();
