@@ -3,7 +3,7 @@
 
 'use strict';
 
-exports.config = {
+var config = {
   // The timeout for each script run on the browser. This should be longer
   // than the maximum time your application needs to stabilize between tasks.
   allScriptsTimeout: 110000,
@@ -12,9 +12,10 @@ exports.config = {
   // with relative paths will be prepended with this.
   baseUrl: 'http://localhost:' + (process.env.PORT || '9000'),
 
-  // If true, only chromedriver will be started, not a standalone selenium.
-  // Tests for browsers other than chrome will not run.
-  chromeOnly: true,
+  // Credientials for Saucelabs
+  sauceUser: process.env.SAUCE_USERNAME,
+
+  sauceKey: process.env.SAUCE_ACCESS_KEY,
 
   // list of files / patterns to load in the browser
   specs: [
@@ -31,7 +32,10 @@ exports.config = {
   // and
   // https://code.google.com/p/selenium/source/browse/javascript/webdriver/capabilities.js
   capabilities: {
-    'browserName': 'chrome'
+    'browserName': 'chrome',
+    'name': 'Fullstack E2E',
+    'tunnel-identifier': process.env.TRAVIS_JOB_NUMBER,
+    'build': process.env.TRAVIS_BUILD_NUMBER
   },
 
   // ----- The test framework -----
@@ -39,12 +43,39 @@ exports.config = {
   // Jasmine and Cucumber are fully supported as a test and assertion framework.
   // Mocha has limited beta support. You will need to include your own
   // assertion framework if working with mocha.
-  framework: 'jasmine',
+  framework: 'mocha',
 
-  // ----- Options to be passed to minijasminenode -----
-  //
-  // See the full list at https://github.com/juliemr/minijasminenode
-  jasmineNodeOpts: {
+  // ----- Options to be passed to mocha -----
+  mochaOpts: {
+    reporter: 'spec',
+    timeout: 30000,
     defaultTimeoutInterval: 30000
+  },
+
+  // Prepare environment for tests
+  params: {
+    serverConfig: require('./server/config/environment')
+  },
+
+  onPrepare: function() {
+    require('babel-core/register');
+    // Load Mocha and Chai + plugins
+    require('./mocha.conf');
+
+    // Expose should assertions (see https://github.com/angular/protractor/issues/633)
+    Object.defineProperty(
+      protractor.promise.Promise.prototype,
+      'should',
+      Object.getOwnPropertyDescriptor(Object.prototype, 'should')
+    );
+
+    var serverConfig = config.params.serverConfig;
+
+    // Setup mongo for tests
+    var mongoose = require('mongoose');
+    mongoose.connect(serverConfig.mongo.uri, serverConfig.mongo.options); // Connect to database
   }
 };
+
+config.params.baseUrl = config.baseUrl;
+exports.config = config;
