@@ -4,29 +4,41 @@
 
 'use strict';
 
-// Set default node environment to development
-process.env.NODE_ENV = process.env.NODE_ENV || 'development';
+import express from 'express';
+import mongoose from 'mongoose';
+mongoose.Promise = require('bluebird');
+import config from './config/environment';
+import http from 'http';
 
-var express = require('express');
-var mongoose = require('mongoose');
-var config = require('./config/environment');
-
-// Connect to database
+// Connect to MongoDB
 mongoose.connect(config.mongo.uri, config.mongo.options);
+mongoose.connection.on('error', function(err) {
+  console.error('MongoDB connection error: ' + err);
+  process.exit(-1);
+});
 
-// Populate DB with sample data
-if(config.seedDB) { require('./config/seed'); }
+// Populate databases with sample data
+if (config.seedDB) { require('./config/seed'); }
 
 // Setup server
 var app = express();
-var server = require('http').createServer(app);
+var server = http.createServer(app);
+var socketio = require('socket.io')(server, {
+  serveClient: config.env !== 'production',
+  path: '/socket.io-client'
+});
+require('./config/socketio')(socketio);
 require('./config/express')(app);
 require('./routes')(app);
 
 // Start server
-server.listen(config.port, config.ip, function () {
-  console.log('Express server listening on %d, in %s mode', config.port, app.get('env'));
-});
+function startServer() {
+  app.angularFullstack = server.listen(config.port, config.ip, function() {
+    console.log('Express server listening on %d, in %s mode', config.port, app.get('env'));
+  });
+}
+
+setImmediate(startServer);
 
 // Expose app
 exports = module.exports = app;
