@@ -12,9 +12,11 @@ module.exports[404] = function pageNotFound(req, res) {
   };
 
   res.status(result.status);
-  res.render(viewFilePath, function (err) {
-    if (err) { return res.json(result, result.status); }
+  res.render(viewFilePath, {}, function(err, html) {
+    if (err) {
+      return res.json(result, result.status);
+    }
 
-    res.render(viewFilePath);
+    res.send(html);
   });
 };
